+++ mainProcess/src/main.ts
@@ -1,10 +1,11 @@
-import {app, BrowserWindow, Menu} from 'electron';
+import { app, BrowserWindow, Menu } from 'electron';
 import { autoUpdater } from "electron-updater";
 
 import path = require("path");
 import "v8-compile-cache";
 
 import "./eventBuses";
+import { BrowserWindowManager } from './controllers/browserWindowManager';
 
 try {
   autoUpdater.checkForUpdatesAndNotify();
@@ -26,6 +27,7 @@ function createWindow() {
     backgroundColor: "#fafafa"
   });
   mainWindow.maximize();
+  BrowserWindowManager.setInstance(mainWindow);
 
   mainWindow.loadFile('./dist-app/index.html');
   // mainWindow.loadURL("http://localhost:8080");
@@ -52,4 +54,3 @@ app.on('activate', () => {
     createWindow();
   }
 });
-