+++ src/components/app/MainMenuComponent.vue
@@ -46,7 +46,9 @@ export default class MainMenuComponent extends Vue {
     this.menus = [
       {title: "General configuration", icon: "mdi-settings", path: "/tests/generalConfig"},
       {title: "Tests", icon: "mdi-bug", path: "/tests/list"},
+      {title: "Engine scripts", icon: "mdi-script", path: "/tests/engineScripts"},
       {title: "Reports", icon: "mdi-clipboard-text", path: "/tests/report"},
+      {title: "Logs", icon: "mdi-file-document", path: "/tests/logs"},
     ];
   }