+++ src/services/fileService.ts
@@ -8,7 +8,7 @@ export class FileService {
     return path.resolve(...paths);
   }
 
-  public static readFile(filePath: string) {
+  public static readFile(filePath: string): Promise<string> {
     return new Promise((resolve, reject) => {
       fs.readFile(filePath, {encoding: "utf-8"}, (err: any, content: string) => {
         if (err) {
@@ -25,6 +25,30 @@ export class FileService {
       fs.writeFile(filePath, content, {encoding: "utf-8"}, (err: any) => {
         if (err) {
           reject("There was en error while saving file: " + err.message);
+        } else {
+          resolve();
+        }
+      });
+    });
+  }
+
+  public static copyFile(originPath: string, destinationPath: string) {
+    return new Promise((resolve, reject) => {
+      fs.copyFile(originPath, destinationPath, (err: any) => {
+        if (err) {
+          reject(err);
+        } else {
+          resolve();
+        }
+      });
+    });
+  }
+
+  public static deleteFile(filePath: string) {
+    return new Promise((resolve, reject) => {
+      fs.unlink(filePath, (err: any) => {
+        if (err) {
+          reject(err);
         } else {
           resolve();
         }