+++ src/store/configurationStore.ts
@@ -43,6 +43,13 @@ export default class ConfigurationStore extends VuexModule {
           FileService.resolvePath([prefixPath, reportPath]) || "";
   }
 
+  public get engineScriptDirectory() {
+    const engineScriptPath = this.currentConfiguration?.paths?.engine_scripts ?? "";
+    const prefixPath = this.configurationPath.substr(0, this.configurationPath.length - "backstop.json".length);
+    return engineScriptPath &&
+          FileService.resolvePath([prefixPath, engineScriptPath]) || "";
+  }
+
   @Mutation
   public addViewport() {
     if (this.currentConfiguration) {
@@ -91,7 +98,7 @@ export default class ConfigurationStore extends VuexModule {
   public setFullConfiguration({ newConfiguration }: any) {
     this.currentConfiguration = newConfiguration;
     this.testsModified = [];
-    if (this .currentConfiguration && Array.isArray(this.currentConfiguration.scenarios)) {
+    if (this.currentConfiguration && Array.isArray(this.currentConfiguration.scenarios)) {
       this.currentConfiguration.scenarios.forEach(() => {
         this.testsModified.push(false);
       });
@@ -275,6 +282,7 @@ export default class ConfigurationStore extends VuexModule {
                   newConfiguration: content
                 });
                 this.context.commit("setPath", path + "/backstop.json");
+                this.context.commit("testLogStore/resetLogs", null, { root: true });
               });
           });
       });
@@ -294,6 +302,7 @@ export default class ConfigurationStore extends VuexModule {
         });
         this.context.commit("setPath", path);
         this.context.commit("updateRecently", path);
+        this.context.commit("testLogStore/resetLogs", null, { root: true });
         BackstopService.setWorkingDir(this.backstopConfigurationDirectory);
         return Promise.resolve();
       });
@@ -313,6 +322,7 @@ export default class ConfigurationStore extends VuexModule {
         });
         this.context.commit("setPath", path);
         this.context.commit("updateRecently", path);
+        this.context.commit("testLogStore/resetLogs", null, { root: true });
         BackstopService.setWorkingDir(this.backstopConfigurationDirectory);
         return Promise.resolve();
       });