+++ src/store/index.ts
@@ -7,6 +7,8 @@ import configurationStore from "./configurationStore";
 import applicationStore from "./applicationStore";
 import testResultStore from "./testResultStore";
 import testRunnerStore from "./testRunnerStore";
+import testLogStore from "./testLogStore";
+import engineScriptStore from "./engineScriptStore";
 
 export default new Vuex.Store({
   state: {
@@ -19,6 +21,8 @@ export default new Vuex.Store({
     configurationStore,
     applicationStore,
     testResultStore,
-    testRunnerStore
+    testRunnerStore,
+    testLogStore,
+    engineScriptStore
   },
 });